Refactor all links to share a superclass
