+++ betty/plugins/link/betty_documentation.py
@@ -4,17 +4,14 @@ The link to Betty's documentation.
 
 from typing import final
 
-from betty.html import NavigationLink
-from betty.link import Link, LinkDefinition
+from betty.link import Link, LinkDefinition, StaticLink
 from betty.locale.localizable.gettext import _
 
 
 @final
 @LinkDefinition(
     "betty-documentation",
-    link=NavigationLink(
-        "https://betty.readthedocs.io/", _("Read the Betty documentation")
-    ),
+    link=StaticLink("https://betty.readthedocs.io/", _("Read the Betty documentation")),
 )
 class BettyDocumentation(Link):
     """
