+++ betty/plugins/link/betty_github.py
@@ -4,17 +4,14 @@ The link to Betty's GitHub.
 
 from typing import final
 
-from betty.html import NavigationLink
-from betty.link import Link, LinkDefinition
+from betty.link import Link, LinkDefinition, StaticLink
 from betty.locale.localizable.gettext import _
 
 
 @final
 @LinkDefinition(
     "betty-github",
-    link=NavigationLink(
-        "https://github.com/bartfeenstra/betty", _("Find Betty on GitHub")
-    ),
+    link=StaticLink("https://github.com/bartfeenstra/betty", _("Find Betty on GitHub")),
 )
 class BettyGithub(Link):
     """
