+++ betty/plugins/link/http_api_doc.py
@@ -4,15 +4,14 @@ The link to the HTTP API documentation.
 
 from typing import final
 
-from betty.html import NavigationLink
-from betty.link import Link, LinkDefinition
+from betty.link import Link, LinkDefinition, StaticLink
 from betty.locale.localizable.gettext import _
 
 
 @final
 @LinkDefinition(
     "http-api-doc",
-    link=NavigationLink("betty-static:///api/index.html", _("API documentation")),
+    link=StaticLink("betty-static:///api/index.html", _("API documentation")),
 )
 class HttpApiDoc(Link):
     """
