+++ betty/tests/test_link.py
@@ -1,7 +1,8 @@
-from betty.html import NavigationLink
-from betty.link import LinkDefinition
+from betty.link import LinkDefinition, StaticLink
+from betty.locale.localizable.plain import Plain
+from betty.locale.localize import DEFAULT_LOCALIZER
 
-_DUMMY_LINK = NavigationLink("https://example.com", "My First Link")
+_DUMMY_LINK = StaticLink("https://example.com", "My First Link")
 
 
 class TestLinkDefinition:
@@ -16,3 +17,15 @@ class TestLinkDefinition:
     def test_primary__with_primary(self) -> None:
         sut = LinkDefinition("my-first-link", link=_DUMMY_LINK, primary=True)
         assert sut.primary
+
+
+class TestStaticLink:
+    def test_url(self) -> None:
+        url = "https://example.com"
+        sut = StaticLink(url, "Hello, world!")
+        assert sut.url.localize(DEFAULT_LOCALIZER) == url
+
+    def test_label(self) -> None:
+        label = Plain("Hello, world!")
+        sut = StaticLink("https://example.com", label)
+        assert sut.label is label
